+++ controllers/book-controller.js
@@ -1,12 +1,15 @@
+
+const issuedBook = require("../dtos/book-dto");
 const { BookModel, UserModel } = require("../models");
+const bookModel = require("../models/book-model");
 
 
-exports.getAllBooks = async(req, res) => {
-    const { id } = req.params;
-    const book = await BookModel.findById(id);
+exports.getAllBooks = async (req, res) => {
+    
+    const books = await BookModel.find();
 
 
-    if(books.lenghth === 0)
+    if(books.length === 0)
     return res.status(404).json({
         success: false,
         message: "No book found",
@@ -36,6 +39,28 @@ exports.getSingleBookById = async (req, res) => {
    });
 };
 
+// Additional route
+exports.getSingleBookByName = async (req, res) => {
+  const { name } = req.params;
+
+  const book = await BookModel.findOne({
+    name: name,
+  });
+
+  if (!book)
+  return res.status(404).json({
+      success: false,
+      message: "Book not found",
+  });
+
+ return res.status(200).json({
+  success: true,
+  data: book,
+
+ });
+};
+
+
 exports.gettAllIssuedBooks = async (req, res) => {
     const users = await UserModel.find({
         issuedBook : {$exists: true},
@@ -53,5 +78,45 @@ exports.gettAllIssuedBooks = async (req, res) => {
     return res.status(200).json({
       success: true,
       data: issuedBooks,
+    });
+  };
+
+
+  exports.addNewBook = async  (req, res) => {
+    const { data } = req.body;
+  
+    if (!data) {
+      return res.status(400).json({
+        success: false,
+        message: "No data provided",
+      });
+    }
+  
+await bookModel.create(data);  
+
+const allBooks = await BookModel.find();
+    
+    return res.status(201).json({
+      success: true,
+      data: allBooks,
+    });
+  };
+
+  exports.updateBookById = async  (req, res) => {
+    const { id } = req.params;
+    const { data } = req.body;
+  
+  const updatedBook = await BookModel.findOneAndUpdate({
+    _id: id,
+  }, 
+  data, 
+  {
+    new: true,
+  }
+  );
+
+    return res.status(200).json({
+      success: true,
+      data: updatedBook,
     });
   };
