+++ controllers/user-controller.js
@@ -1,3 +1,181 @@
-const { BookModel, UserModel } = require("../models");
+const { UserModel,BookModel} = require('../models');
+const userModel = require('../models/user-model');
 
 
+
+exports.getAllUsers = async (req,res) => {
+    const users =  await UserModel.find();
+
+    if(users.length === 0){
+        return res.status(404).json({
+            success: false,
+            message: "No Users found",
+        });
+    }
+
+
+    return res.status(200).json({
+        success: true,
+        data: users,
+        
+    });
+};
+
+exports.getSingleUserById = async (req,res) =>{
+    const {id} = req.params
+    const user = await UserModel.findById({_id: id});
+    if (!user) {
+        return res.status(404).json({
+        success:false,
+        message:"User not found",
+        });
+    }
+    return res.status(200).json({
+        success: true,
+        data: user,
+        
+    });
+};
+
+exports.createNewUser = async (req,res) => {
+    const { data } = req.body;
+  
+    const newUser = await UserModel.create(data);
+
+    
+
+    
+    return res.status(201).json({
+        success: true,
+        data: newUser,
+
+    });
+
+};
+
+
+exports.updateUserById = async (req,res) =>{
+    const {id} = req.params;
+    const {data} = req.body;
+
+const updatedUserData = await UserModel.findOneAndUpdate(
+ {
+    _id: id,
+
+}, 
+{
+    $set: {
+        ...data,
+    },
+},
+{
+    new: true,
+}
+
+);
+    
+
+
+return res.status(200).json({
+    success: true,
+    data: updatedUserData,
+});
+};
+
+exports.deleteUser = async (req,res) => {
+    const {id} = req.params;
+    
+    const user = await UserModel.deleteOne({
+        _id: id,
+    });
+    if(!user){
+        return res.status(202).json({
+            success: false,
+            message: "User to be deleted was not found",
+
+        });
+
+    }
+
+   
+
+    return res.status(200).json({ success: true, message: "deleted the user successfully"});
+
+};
+
+exports.getSubscriptionDetailsById = async (req,res) => {
+    const { id } = req.params;
+
+    const user = await userModel.findById(id);
+
+    if(!user) 
+    return res.status(404).json({
+        success: false,
+        message: "User not found",
+    });
+ 
+    const getDateInDays = (data = "") => {
+        let date;
+        if(data === ""){
+
+            // current date
+            date = new Date();
+        } else {
+
+            // getting date on basis of data variable
+            date = new Date(data);
+        }
+        let days = Math.floor(date / (1000*60*60*24));
+        return days;
+    };
+
+    const subscriptionType = (date) => {
+        if(user.subscriptionType === "Basic"){
+            date = date + 90;
+
+        } else if (user.subscriptionType === "standard"){
+            date = date + 180;
+
+        } else if (user.subscriptionType === "premium"){
+            date = date + 365;
+        }
+
+        return date;
+
+    };
+
+   // subscription expiration calculation
+   // january 1, 1970, UTC. //milliseconds
+   let returnDate = getDateInDays(user.returnDate);
+   let currentDate = getDateInDays();
+   let subscriptionDate = getDateInDays(user.subscriptionDate);
+   let subscriptionExpiration = subscriptionType(subscriptionDate);
+
+
+
+   const data = {
+    ...user._doc,
+    subscriptionExpired: subscriptionExpiration < currentDate,
+    daysLeftForExpiration: 
+     subscriptionExpiration <= currentDate 
+     ? 0 
+     : subscriptionExpiration - currentDate,
+    
+     
+    fine:
+      returnDate < currentDate ?
+      subscriptionExpiration <= currentDate
+      ? 200 
+      : 100
+     : 0,
+
+
+   };
+
+   res.status(200).json({
+    success: true,
+    data,
+
+   });
+    
+};
