+++ dtos/book-dto.js
