+++ models/user-model.js
@@ -23,6 +23,11 @@ const userSchema = new Schema(
         ref: "Book",
         required: false,
        },
+       issuedDate: {
+         type: String,
+         required: false,
+        },
+ 
 
        returnDate: {
         type: String,
