+++ routes/books.js
@@ -7,7 +7,7 @@ const { users } = require("../data/users.json");
 
 const { UserModel, BookModel } = require("../models");
 
-const { getAllBooks, getSingleBookById, gettAllIssuedBooks } = require("../controllers/book-controller");
+const { getAllBooks, getSingleBookById, gettAllIssuedBooks, addNewBook, updateBookById, getSingleBookByName } = require("../controllers/book-controller");
 
 const router = express.Router();
 
@@ -31,6 +31,8 @@ router.get("/",getAllBooks);
 
 router.get("/:id", getSingleBookById);
 
+router.get("/getbook/name/:name", getSingleBookByName);
+
 /**
  * Route: /books/issued/by-user
  * Method: GET
@@ -39,6 +41,8 @@ router.get("/:id", getSingleBookById);
  * parameters: none
  */
 router.get("/issued/by-user", gettAllIssuedBooks);
+
+
 
 /**
  * Route: /books
@@ -49,32 +53,7 @@ router.get("/issued/by-user", gettAllIssuedBooks);
  * data: author,name,genre,price,publisher,id
  */
 
-router.post("/", (req, res) => {
-  const { data } = req.body;
-
-  if (!data) {
-    return res.status(400).json({
-      success: false,
-      message: "No data provided",
-    });
-  }
-
-  const book = books.find((each) => each.id === data.id);
-
-  if (book) {
-    return res.status(404).json({
-      success: false,
-      message: "book already exist with this id , please use a unique id",
-    });
-  }
-
-  const allbooks = [...books, data];
-
-  return res.status(201).json({
-    success: true,
-    data: allbooks,
-  });
-});
+router.post("/",addNewBook);
 
 /**
  * Route: /books/:id
@@ -85,30 +64,7 @@ router.post("/", (req, res) => {
  * data: author,name,genre,price,publisher,id
  */
 
-router.put("/:id", (req, res) => {
-  const { id } = req.params;
-  const { data } = req.body;
-
-  const book = books.find((each) => each.id === id);
-
-  if (!book) {
-    return res.status(400).json({
-      success: false,
-      message: "book not found with this particular id",
-    });
-  }
-
-  const updateData = books.map((each) => {
-    if (each.id === id) {
-      return { ...each, ...data };
-    }
-    return each;
-  });
-  return res.status(200).json({
-    success: true,
-    data: updateData,
-  });
-});
+router.put("/:id",updateBookById);
 
 /**
  * Route: /books/issued/with-fine
