+++ routes/users.js
@@ -1,4 +1,5 @@
 const express = require("express");
+const { getAllUsers, getSingleUserById, deleteUser, updateUserById, createNewUser, getSubscriptionDetailsById } = require("../controllers/user-controller");
 
 const {users} = require('../data/users.json');
 
@@ -11,21 +12,10 @@ const router = express.Router();
  * Access: Public
  * parameters: None
  */
- router.get('/users/:id',(req,res) =>{
-    const {id} = req.params
-    const user = users.find((each) => each.id === id  );
-    if (!user) {
-        return res.status(404).json({
-        success:false,
-        message:"User not found",
-        });
-    }
-    return res.status(200).json({
-        success: true,
-        data: user,
-        
-    });
-});
+
+ router.get('/', getAllUsers );
+
+
 
 /**
  * Route: /users/:id
@@ -35,12 +25,8 @@ const router = express.Router();
  * parameters: None
  */
 
-router.get('/users', (req,res) =>{
-    res.status(200).json({
-        success: true,
-        data: users,
-    });
-});
+ router.get('/:id', getSingleUserById);
+
 
 /**
  * Route: /users
@@ -49,36 +35,7 @@ router.get('/users', (req,res) =>{
  * Access: Public
  * parameters: None
  */
-router.post('/users',(req,res) => {
-    const {id,name,surname,email,subscriptionType,subscriptionDate} = 
-    req.body;
-
-    const user = users.find((each) => each.id === id);
-
-    if(user){
-        return res.status(404).json({
-            success: false,
-            message: "User exists with this id",
-
-        });
-
-    }
-
-    users.push({
-        id,
-        name,
-        surname,
-        email,
-        subscriptionDate,
-        subscriptionType,
-    });
-    return res.status(201).json({
-        success: true,
-        data: users,
-
-    });
-
-});
+router.post('/',createNewUser);
 
 /**
  * Route: /users/:id
@@ -87,34 +44,10 @@ router.post('/users',(req,res) => {
  * Access: Public
  * parameters: id
  */
-router.put('/users/:id', (req,res) =>{
-    const {id} = req.params;
-    const {data} = req.body;
-
-    const user = users.find((each) => each.id === id);
-
-    if(!user) 
-    return res.status(404).json({success: false, message: "User not found"});
-
-const updatedUser = users.map((each) => {
-    if(each.id === id){
-        return {
-            ...each,
-            ...data,
-        };
-
-    }
-
-    return each;
-});
-
-return res.status(200).json({
-    success: true,
-    data: updatedUser,
-});
+router.put('/:id', updateUserById);
 
 
-});
+
 
 
 /**
@@ -125,25 +58,7 @@ return res.status(200).json({
  * parameters: id
  */
 
-router.delete('/users/:id', (req,res) => {
-    const {id} = req.params;
-    const user = users.find((each) => each.id === id);
-
-    if(!user){
-        return res.status(202).json({
-            success: false,
-            message: "User to be deleted was not found",
-
-        });
-
-    }
-
-    const index = users.indexOf(user);
-    users.splice(index,1);
-
-    return res.status(200).json({ success: true, data: users});
-
-});
+router.delete('/:id', deleteUser);
 /**
  * Route: /users/subscription-details/:id
  * Method: GET
@@ -152,81 +67,6 @@ router.delete('/users/:id', (req,res) => {
  * parameters: id
  */
 
-router.get('/subscription-details/:id', (req,res) => {
-    const { id } = req.params;
-
-    const user = users.find((each) => each.id === id);
-
-    if(!user) 
-    return res.status(404).json({
-        success: false,
-        message: "User not found",
-    });
- 
-    const getDateInDays = (data = "") => {
-        let date;
-        if(data === ""){
-
-            // current date
-            date = new Date();
-        } else {
-
-            // getting date on basis of data variable
-            date = new Date(data);
-        }
-        let days = Math.floor(date / (1000*60*60*24));
-        return days;
-    };
-
-    const subscriptionType = (date) => {
-        if(user.subscriptionType === "Basic"){
-            date = date + 90;
-
-        } else if (user.subscriptionType === "standard"){
-            date = date + 180;
-
-        } else if (user.subscriptionType === "premium"){
-            date = date + 365;
-        }
-
-        return date;
-
-    };
-
-   // subscription expiration calculation
-   // january 1, 1970, UTC. //milliseconds
-   let returnDate = getDateInDays(user.returnDate);
-   let currentDate = getDateInDays();
-   let subscriptionDate = getDateInDays(user.subscriptionDate);
-   let subscriptionExpiration = subscriptionType(subscriptionDate);
-
-
-
-   const data = {
-    ...user,
-    subscriptionExpired: subscriptionExpiration < currentDate,
-    daysLeftForExpiration: 
-     subscriptionExpiration <= currentDate 
-     ? 0 
-     : subscriptionExpiration - currentDate,
-    
-     
-    fine:
-      returnDate < currentDate ?
-      subscriptionExpiration <= currentDate
-      ? 200 
-      : 100
-     : 0,
-
-
-   };
-
-   res.status(200).json({
-    success: true,
-    data,
-
-   });
-    
-});
+router.get('/subscription-details/:id', getSubscriptionDetailsById);
 
 module.exports = router;
